Reworks and debugs the CRT functions
Instead of just returning the value x which solves the given system of simultaneous congruences, the function CRT and its implementations in PolynomialSimultaneousCongruencesProblem and IntegerSimultaneousCongruencesProblem now return an array with two elements with the first element being the value x, which solves the system, and the second element being the new modulus of the congruence, which is basically the lcm of all moduli of the system. JavaDocs, FFaplAPISpec.bin, sunset.properties files and example ffapl files have been updated accordingly.

Also, a bug has been fixed in IntegerSimultaneousCongruencesProblem and PolynomialSimultaneousCongruencesProblem which had the effect that the solution for the system of simultaneous congruences was larger than the module. To fix this, before returning the result, x is reduced modulo the module again.