Pass configuration into the interpreter through a properties object
Pass in runtime configuration (isomorphism calculation strategy and
timeout for now) in a new `FFaplRuntimeProperties` object.

This reduces the coupling between the `ffapl` and `sunset` packages.
The user-configurable options for the isomorphism calculation are now
handled in the Sunset application and simply passed in when starting the
interpreter.

This design allows for future extensions with additional options while
keeping the existing interface.