Configure Syntax-Matcher colors explicitly
Reduce the binding between the `ffapl` and `sunset` package by removing
the implicit configuration of the `*Matcher` instances in the `ffapl`
package.

The matchers now come with a baseline configuration and are properly
set-up in the Sunset application to match the desired styling (i.e. font,
color, etc.) for syntax highlighting.