Add TeaVM shim to call the FFaplInterpreter from JavaScript
Provides an entry point for TeaVM which manages the `execute()`
interface to run an FFapl program from JavaScript as well as a wrapper
to call a `log()` implementation in JavaScript.

Provides a shim to a minimal set of `java.security.MessageDigest` and
`javax.crypto.Cipher` backed by a JavaScript implementation of
`CryptoProvider`.